Fixed bugs

+++ src/main/java/com/germanium/lmsuserservice/config/JwtTokenUtil.java
@@ -1,7 +1,6 @@
 package com.germanium.lmsuserservice.config;
 
 import java.io.Serializable;
-import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
@@ -20,7 +19,7 @@ public class JwtTokenUtil implements Serializable {
 
 	private static final long serialVersionUID = -2550185165626007488L;
 
-	public static final long JWT_TOKEN_VALIDITY = 5*60*60;
+	public static final long JWT_TOKEN_VALIDITY = Long.valueOf(5*60*60);
 
 	@Value("${jwt.secret}")
 	private String secret;
